Refactor imports to use sigic_auth namespace for Keycloak authentication
