+++ src/sigic_geonode/upload/views.py
@@ -2,7 +2,7 @@ from importer.api.views import ImporterViewSet
 from geonode.documents.api.views import DocumentViewSet
 from rest_framework.authentication import BasicAuthentication, SessionAuthentication
 from oauth2_provider.contrib.rest_framework import OAuth2Authentication
-from sigic_geonode.auth.keycloak import KeycloakJWTAuthentication
+from sigic_geonode.sigic_auth.keycloak import KeycloakJWTAuthentication
 
 # 
 
@@ -20,7 +20,7 @@ class SigicDocumentViewSet(DocumentViewSet):
     def get_authenticators(self):
         from rest_framework.authentication import BasicAuthentication, SessionAuthentication
         from oauth2_provider.contrib.rest_framework import OAuth2Authentication
-        from sigic_geonode.auth.keycloak import KeycloakJWTAuthentication
+        from sigic_geonode.sigic_auth.keycloak import KeycloakJWTAuthentication
 
         print("✅ get_authenticators personalizado en uso")
